Fix l4linux block driver for large partitions
Fixes #678

+++ ports-foc/include/genode/block.h
@@ -40,7 +40,7 @@ L4_CV void* genode_block_request(unsigned idx, unsigned long sz,
                                  void *req, unsigned long *offset);
 
 L4_CV void genode_block_submit(unsigned idx, unsigned long queue_offset,
-                               unsigned long size, unsigned long disc_offset,
+                               unsigned long size, unsigned long long disc_offset,
                                int write);
 
 L4_CV void genode_block_collect_responses(unsigned idx);
